Select Iterator

+++ src/offset.rs
@@ -1,7 +1,7 @@
 use std::io::SeekFrom;
 
 #[derive(Debug, Default, Copy, Clone)]
-pub struct Offset(u64);
+pub struct Offset(i64);
 
 impl Offset {
     pub fn as_usize(&self) -> usize {
@@ -11,19 +11,19 @@ impl Offset {
 
 impl From<u16> for Offset {
     fn from(p: u16) -> Self {
-        Self(p as u64)
+        Self(p as i64)
     }
 }
 
-impl From<u64> for Offset {
-    fn from(p: u64) -> Self {
+impl From<i64> for Offset {
+    fn from(p: i64) -> Self {
         Self(p)
     }
 }
 
 impl Into<SeekFrom> for Offset {
     fn into(self) -> SeekFrom {
-        SeekFrom::Start(self.0)
+        SeekFrom::Start(self.0 as u64)
     }
 }
